Changed image to use raw url
build.cfengine.com can pulls in broken images if I use the relative path to an
image. Using the raw url works for both rendering on github as well as for
build.cfengine.com (but, sadly, it prevents me from seeing the image in emacs,
oh well, such is life).